requirements.txt
accelerate==0.26.1
bitsandbytes==0.42.0
datasets==2.17.1
diffusers==0.13.0
einops==0.7.0
huggingface_hub==0.20.3
matplotlib==3.5.2
numpy==1.24.4
Pillow==10.2.0
scikit_image==0.19.3
seaborn==0.13.2
thop==0.1.0.post2207010342
timm==0.6.5
torch==2.1.0+cu121
torch_geometric==2.5.0
torchmetrics==1.3.0.post0
torchvision==0.16.0+cu121
tqdm==4.66.1
transformers==4.37.1
wandb==0.16.2
